+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## 28.02.2023
-    - `feature`: Установка ts-loader
+    - `feature`: Установка react
 
     - `feature`: Настройка конфигурации webpack.config.js
 
     - `feature`: Установка реакта и необходимых библиотек для ts
         npm i react react-dom
         npm i --save-dev @types/react @types/react-dom
+    - `feature`: Установка ts-loader
 
     - `feature`: npm i --save-dev typescript ts-node @types/node - Установка typescript и добавление конфигурационных файлов