+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 28.02.2023
+    - `feature`: Установка плагинов для сборки
+        npm i webpack html-webpack-plugin clean-webpack-plugin --save-dev
     - `feature`: Установка react
 
     - `feature`: Настройка конфигурации webpack.config.js