+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 29.02.2023
+    - `feature`: Добавлены некоторые файлы html и tsx
+
     - `feature`: Установлен webpack-dev-server
         npm i -D webpack-dev-server
 
@@ -18,11 +20,14 @@
     - `feature`: Установлен реакт и необходимые библиотеки для ts
         npm i react react-dom
         npm i --save-dev @types/react @types/react-dom
+
     - `feature`: Установлен ts-loader
 
-    - `feature`: npm i --save-dev typescript ts-node @types/node - Установлен typescript и добавлен конфигурационный файл
+    - `feature`: Установлен typescript и добавлен конфигурационный файл
+        npm i --save-dev typescript ts-node @types/node
 
-    - `feature`: npm i -D webpack webpack-cli - Установлен webpack
+    - `feature`: Установлен webpack
+        npm i -D webpack webpack-cli
         `-D - devDependencies` - только для разработки
 
     - `feature`: Добавлен package.json и CHANGELOG